+++ src/main/kotlin/io/github/jamalam/database/entity/ServerConfig.kt
@@ -28,7 +28,8 @@ data class ServerConfig(
     var moderationConfig: ServerModerationConfig,
     var tagsConfig: ServerTagsConfig,
     var notificationsConfig: ServerNotificationsConfig,
-    var filePasteConfig: ServerFilePasteConfig
+    var filePasteConfig: ServerFilePasteConfig,
+    var phishingConfig: ServerPhishingConfig,
 )
 
 data class ServerQuotesConfig(
@@ -63,4 +64,8 @@ data class ServerNotificationsConfig(
 data class ServerFilePasteConfig(
     var enabled: Boolean,
     var hastebinUrl: String
+)
+
+data class ServerPhishingConfig(
+    var enabled: Boolean
 )