+++ src/main/kotlin/io/github/jamalam/extensions/moderation/ModerationExtension.kt
@@ -186,12 +186,14 @@ class ModerationExtension : Extension() {
                                         userAuthor(user.asUser())
                                         now()
                                         error()
-                                        stringField("Duration", arguments.duration.toString())
+                                        stringField("Duration", arguments.duration.toPrettyString())
                                         stringField("Reason", arguments.reason)
                                     }
                                 }
+                            }
 
-                                scheduler.schedule(arguments.duration.seconds.toLong()) {
+                            scheduler.schedule(arguments.duration.seconds.toLong()) {
+                                if (!member.isBot) {
                                     member.dm {
                                         embed {
                                             info("Unmuted in ${guild!!.asGuild().name}!")
@@ -200,25 +202,26 @@ class ModerationExtension : Extension() {
                                             success()
                                         }
                                     }
+                                }
 
-                                    guild?.getLogChannel()?.createEmbed {
-                                        info("Member Unmuted")
-                                        userAuthor(user.asUser())
-                                        log()
-                                        userField("Member", member.asUser())
-                                    }
+                                guild?.getLogChannel()?.createEmbed {
+                                    info("Member Unmuted")
+                                    userAuthor(user.asUser())
+                                    log()
+                                    userField("Member", member.asUser())
+                                }
 
-                                    guild?.getPublicModLogChannel()?.createEmbed {
-                                        info("Member Unmuted")
-                                        userAuthor(user.asUser())
-                                        log()
-                                        userField("Member", member.asUser())
-                                    }
+                                guild?.getPublicModLogChannel()?.createEmbed {
+                                    info("Member Unmuted")
+                                    userAuthor(user.asUser())
+                                    log()
+                                    userField("Member", member.asUser())
                                 }
                             }
 
                             member.edit {
-                                timeoutUntil = Clock.System.now().plus(arguments.duration, TimeZone.currentSystemDefault())
+                                timeoutUntil =
+                                    Clock.System.now().plus(arguments.duration, TimeZone.currentSystemDefault())
 
                                 reason =
                                     "Muted by ${user.asUser().username} for ${arguments.duration} with the reason '${arguments.reason}'"
@@ -230,7 +233,7 @@ class ModerationExtension : Extension() {
                                 now()
                                 log()
                                 userField("Member", member.asUser())
-                                stringField("Duration", arguments.duration.toString())
+                                stringField("Duration", arguments.duration.toPrettyString())
                                 stringField("Reason", arguments.reason)
                             }
 
@@ -240,7 +243,7 @@ class ModerationExtension : Extension() {
                                 now()
                                 log()
                                 userField("Member", member.asUser())
-                                stringField("Duration", arguments.duration.toString())
+                                stringField("Duration", arguments.duration.toPrettyString())
                                 stringField("Reason", arguments.reason)
                             }